Debugged 2 and added alerts

+++ src/Components/TodoList.jsx
@@ -1,4 +1,5 @@
 import React, { useEffect, useState } from "react";
+import Swal from "sweetalert2";
 import "./TodoList.css";
 import Button from "@mui/material/Button";
 import AddIcon from "@mui/icons-material/Add";
@@ -7,22 +8,30 @@ import Listing from "./Listing";
 export default function TodoList() {
   //Function to fetch the local storage
   const getItem = () => {
-    let list = localStorage.getItem('newItem');
-    console.log('list:', list);
+    let list = localStorage.getItem("newItem");
+    console.log("list:", list);
 
-    if(list) return JSON.parse(localStorage.getItem('newItem'));
+    if (list) return JSON.parse(localStorage.getItem("newItem"));
     else return [];
-  }
-  
+  };
+
   const [item, setItem] = useState("");
   const [newItem, setNewItem] = useState(getItem());
+  const [bool, setBool] = useState(false);
 
-  
   const addItem = () => {
+    setBool(false);
     //To save the previous inputs and also add new inputs to the array
     setNewItem((previous) => {
       if (item.trim() !== "") {
         setItem("");
+        Swal.fire({
+          position: "top-end",
+          icon: "success",
+          title: "Your work has been saved",
+          showConfirmButton: false,
+          timer: 1500,
+        });
         return [...previous, { text: item, id: new Date().getTime() }];
       } else {
         setItem("");
@@ -44,14 +53,42 @@ export default function TodoList() {
     setNewItem([]);
   };
 
-  const updateItems = (index) => {
-    //Deleting item from list
-    setNewItem(() => {
-      return newItem.filter((val) => {
-          if(val.id === index) return setItem(val.text);
-        return val.id !== index;
-      });
+  const updateAlert = () => {
+    Swal.fire({
+      title: "Do you want to save the changes?",
+      showDenyButton: true,
+      showCancelButton: true,
+      confirmButtonText: "Save",
+      denyButtonText: `Don't save`,
+    }).then((result) => {
+      /* Read more about isConfirmed, isDenied below */
+      if (result.isConfirmed) {
+        addItem();
+        Swal.fire("Saved!", "", "success");
+      } else if (result.isDenied) {
+        Swal.fire("Changes are not saved", "", "info");
+      }
     });
+  };
+
+  const updateItems = (index) => {
+    setBool(true);
+    //Deleting item from list
+    if (!bool) {
+      setNewItem(() => {
+        return newItem.filter((val) => {
+          if (val.id === index) return setItem(val.text);
+          return val.id !== index;
+        });
+      });
+    } else {
+      Swal.fire({
+        icon: "error",
+        title: "Oops...",
+        text: "Already in Update!",
+        // footer: '<a href="">Why do I have this issue?</a>'
+      });
+    }
   };
 
   return (
@@ -72,7 +109,7 @@ export default function TodoList() {
               value={item}
               onChange={addEvent}
             />
-            <Button className="myBtn" onClick={addItem}>
+            <Button className="myBtn" onClick={bool ? updateAlert : addItem}>
               <AddIcon />
             </Button>
